Tech-debt: Move auto-allocations (and App) source to new structure
