+++ src/AutoAllocation/tests/manualAllocation.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import { ManualAllocation } from "./index";
+import { ManualAllocation } from "../components/ManualAllocation";
 
 Enzyme.configure({ adapter: new Adapter() });
