+++ src/AutoAllocation/tests/unallocated.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Unallocated from './index';
+import Unallocated from '../components/Unallocated';
 
 Enzyme.configure({ adapter: new Adapter() });
