+++ src/KeyworkerProfile/index.scss
@@ -1,5 +1,4 @@
 @import "colours/palette";
-@import "colours/palette";
 
 .searchForm {
   background-color: $grey-4;
