+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
-import { AppContainer } from './App';
+import { AppContainer } from './KeyworkerManagement/App';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
 import allocationApp from './redux/reducers';
